requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
lxml==4.3.3
numpy==1.16.3
openpyxl==2.6.2
Pillow==6.0.0
PyAutoGUI==0.9.42
PyGetWindow==0.0.4
PyJWT==1.7.1
PyMsgBox==1.0.6
PyPDF2==1.26.0
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.20
PySocks==1.6.8
python-docx==0.8.10
PyTweening==1.0.3
pytz==2019.1
requests==2.21.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
twilio==6.26.3
urllib3==1.24.2
